TLZ-15 Wrong type TINYINT for Status field at logDbH2.sql script

+++ src/main/resources/sql/logDbH2.sql
@@ -6,7 +6,7 @@ CREATE TABLE IF NOT EXISTS log (
   log_request_uri CLOB NOT NULL ,
   log_request_body CLOB NOT NULL ,
   log_response_body CLOB NOT NULL ,
-  log_response_status TINYINT NOT NULL ,
+  log_response_status SMALLINT NOT NULL ,
   log_stacktrace CLOB NULL
 );
